refactor: replaced 'PcodeShape' with 'Bubble'
former 'PcodeShape' component is now the 'Bubble' component; introducing two variants of the shape: "centered" and "shifted"
