+++ components/Menu.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Box, Flex, Link } from 'rebass';
-import { PcodeShape } from './PcodeShape';
+import { Bubble } from './Bubble';
 import { Color } from '../Color.enum';
 import { CSSObject } from '@styled-system/css';
 import styled from 'styled-components';
@@ -81,7 +81,7 @@ export const Menu: React.FC<Props> = ({
                             </Flex>
                         </LinkLayout>
                         <PcodeShapeLayout>
-                            <PcodeShape width="80vh" color={Color.Primary} />
+                            <Bubble width="80vh" color={Color.Primary} />
                         </PcodeShapeLayout>
                     </ShapeAndLinksLayout>
                 </Flex>
